% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingSP_ldscMR.R
\name{gettingSP_ldscMR}
\alias{gettingSP_ldscMR}
\title{Title}
\usage{
gettingSP_ldscMR(
  input.df,
  trait.names,
  run_ldsc = TRUE,
  run_MR = TRUE,
  saveRFiles = TRUE,
  hm3,
  ld
)
}
\arguments{
\item{ld}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
