% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sumstats.R
\name{merge_sumstats}
\alias{merge_sumstats}
\title{Title}
\usage{
merge_sumstats(
  input.files,
  trait.names,
  LD.filepath,
  rho.filepath,
  mafT = 0.005,
  infoT = 0.99
)
}
\arguments{
\item{input.files}{\itemize{
\item list of data frames, where each data frame contains the summary statistics of a trait to use
}}

\item{trait.names}{\itemize{
\item vector containing the trait names in the order they're found in 'input files'
}}

\item{mafT}{\itemize{
\item Minor allele frequency threshold of selection
}}

\item{infoT}{\itemize{
\item SNP imputation quality threshold
}}

\item{LD.file}{\itemize{
\item LD scores file, either obtained from Alkes group (1000G) or the one provided (UK10K)
}}

\item{rho.file}{\itemize{
\item Genotyped SNP-specific LD scores
}}
}
\value{

}
\description{
Title
}
