% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_sumstats.R
\name{munge_sumstats}
\alias{munge_sumstats}
\title{Title - Read the columns from the summary statistics files and edit them as required}
\usage{
munge_sumstats(input.files, trait.names)
}
\arguments{
\item{input.files}{\itemize{
\item list of data frames, where each data frame contains the summary statistics of a trait to use
}}

\item{trait.names}{\itemize{
\item vector containing the trait names in the order they're found in 'input files'
}}
}
\value{
\itemize{
\item returns the same list of data frames, but with updated column names, begin.time and log.file
}
}
\description{
Title - Read the columns from the summary statistics files and edit them as required
}
\keyword{internal}
