% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairTrait_singleStep_likelihood.R
\name{pairTrait_singleStep_likelihood}
\alias{pairTrait_singleStep_likelihood}
\title{Title}
\usage{
pairTrait_singleStep_likelihood(
  theta,
  betXY,
  pi1,
  sig1,
  w8s,
  m0,
  pi_U = 0.1,
  i_X,
  i_Y,
  nX,
  nY,
  model = "comp",
  bn = 2^8,
  bins = 15,
  M = 1e+07
)
}
\arguments{
\item{M}{}
}
\value{
NOT EXPORTED @export
}
\description{
Title
}
\keyword{internal}
