% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairTrait_twoStep_likelihood.R
\name{pairTrait_twoStep_likelihood}
\alias{pairTrait_twoStep_likelihood}
\title{Title}
\usage{
pairTrait_twoStep_likelihood(
  theta,
  betXY,
  pi1,
  sig1,
  w8s,
  m0,
  pi_U = 0.1,
  pi_X,
  pi_Y,
  i_X,
  i_Y,
  nX,
  nY,
  model = "comp",
  bn = 2^8,
  bins = 15,
  M = 1e+07
)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
Title
}
