% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleTrait_likelihood.R
\name{singleTrait_likelihood}
\alias{singleTrait_likelihood}
\title{Title}
\usage{
singleTrait_likelihood(
  theta,
  betX,
  pi1,
  sig1,
  weights,
  m0,
  nX,
  bn = 2^8,
  bins = 15,
  M = 1e+07
)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
Title
}
