% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockJK_singleStep_lhc_mr.R
\name{slurm_blockJK_singleStep_lhc_likelihood}
\alias{slurm_blockJK_singleStep_lhc_likelihood}
\title{Title}
\usage{
slurm_blockJK_singleStep_lhc_likelihood(par, start_ind, end_ind)
}
\arguments{
\item{end_ind}{}
}
\value{

}
\description{
Title
}
