% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_blockJK_singleStep_likelihood.R
\name{slurm_blockJK_singleStep_likelihood}
\alias{slurm_blockJK_singleStep_likelihood}
\title{Title}
\usage{
slurm_blockJK_singleStep_likelihood(par, start_ind, end_ind)
}
\arguments{
\item{par}{pp}

\item{end_ind}{}
}
\value{
NOT EXPORTED @export
}
\description{
Title
}
