% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_pairTrait_twoStep_likelihood.R
\name{slurm_pairTrait_twoStep_likelihood}
\alias{slurm_pairTrait_twoStep_likelihood}
\title{Title}
\usage{
slurm_pairTrait_twoStep_likelihood(par)
}
\arguments{
\item{par}{}
}
\value{
NOT EXPORTED @export
}
\description{
Title
}
\keyword{internal}
