% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_singleTrait_likelihood.R
\name{slurm_singleTrait_likelihood}
\alias{slurm_singleTrait_likelihood}
\title{Title}
\usage{
slurm_singleTrait_likelihood(par)
}
\arguments{
\item{par}{}
}
\value{

}
\description{
Title
}
