% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SP.R
\name{calculate_SP}
\alias{calculate_SP}
\title{Title}
\usage{
calculate_SP(
  input.df,
  trait.names,
  log.file = NA,
  run_ldsc = TRUE,
  run_MR = TRUE,
  saveRFiles = TRUE,
  hm3 = NA,
  ld = NA,
  nStep = 2,
  SP_single = 3,
  SP_pair = 50,
  SNP_filter = 10,
  nCores = NA,
  M = 1e+07
)
}
\arguments{
\item{jobs}{}
}
\value{

}
\description{
Title
}
