% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_mr.R
\name{lhc_mr}
\alias{lhc_mr}
\title{Title}
\usage{
lhc_mr(
  input.df_filtered,
  trait.names,
  SP_matrix,
  iX,
  iY,
  piX = NA,
  piY = NA,
  SP_pair = 100,
  partition = NA,
  StepNum = 2,
  paral_method = "rslurm",
  nBlock = 200,
  M = 1e+07
)
}
\arguments{
\item{nBlock}{}
}
\value{

}
\description{
Title
}
