% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_LDfiles.R
\name{munge_LDfiles}
\alias{munge_LDfiles}
\title{Title}
\usage{
munge_LDfiles(input.files, trait.names, log.name)
}
\arguments{
\item{input.files}{\itemize{
\item list of data frames, where each data frame contains the summary statistics of a trait to use
}}

\item{trait.names}{\itemize{
\item vector containing the trait names in the order they're found in 'input files'
}}

\item{log.name}{\itemize{
\item the name of the log.file connection to append log descriptions
}}
}
\value{
\itemize{
\item returns the same list of data frames, but with updated column names, and log.name
}
}
\description{
Title
}
\keyword{internal}
